refactor: aggregate some method and use null instead of undefined
