+++ packages/blocks/src/page-block/edgeless/connector-manager.ts
@@ -1,20 +1,20 @@
-import type {
-  Connection,
-  ConnectorElement,
-  SurfaceManager,
-} from '@blocksuite/phasor';
 import {
   almostEqual,
   AStarRunner,
   Bound,
   clamp,
+  type Connection,
+  type ConnectorElement,
   ConnectorMode,
   getBoundFromPoints,
   type IConnector,
   type IVec,
+  lineIntersects,
+  linePolygonIntersects,
   Overlay,
   polygonNearestPoint,
   sign,
+  type SurfaceManager,
   Vec,
 } from '@blocksuite/phasor';
 import { assertEquals, assertExists } from '@blocksuite/store';
@@ -24,12 +24,444 @@ import {
   type TopLevelBlockModel,
 } from '../../__internal__/utils/types.js';
 import type { EdgelessPageBlockComponent } from './edgeless-page-block.js';
-import {
-  computePoints,
-  getAnchors,
-  getConnectableNearestAnchor,
-} from './utils/connector.js';
 import { isTopLevelBlock } from './utils/query.js';
+import type { Selectable } from './utils/selection-manager.js';
+
+export function isConnectorAndBindingsAllSelected(
+  connector: ConnectorElement,
+  selected: Selectable[]
+) {
+  const connectorSelected = selected.find(s => s.id === connector.id);
+  if (!connectorSelected) {
+    return false;
+  }
+  const { source, target } = connector;
+  const startSelected = selected.find(s => s.id === source?.id);
+  const endSelected = selected.find(s => s.id === target?.id);
+  if (!source.id && !target.id) {
+    return true;
+  }
+  if (!source.id && endSelected) {
+    return true;
+  }
+  if (!target.id && startSelected) {
+    return true;
+  }
+  if (startSelected && endSelected) {
+    return true;
+  }
+  return false;
+}
+
+export function getAnchors(ele: Connectable) {
+  const b = Bound.deserialize(ele.xywh);
+  const offset = 10;
+  const anchors: { point: IVec; coord: IVec }[] = [];
+  const coords = [
+    [0.5, 0],
+    [0.5, 1],
+    [0, 0.5],
+    [1, 0.5],
+  ];
+  [
+    [b.center[0], b.y - offset],
+    [b.center[0], b.maxY + offset],
+    [b.x - offset, b.center[1]],
+    [b.maxX + offset, b.center[1]],
+  ].forEach((vec, index) => {
+    const rst = connectableIntersectLine(ele, [b.center, vec]);
+    if (rst) anchors.push({ point: rst[0], coord: coords[index] });
+  });
+  return anchors;
+}
+
+export function connectableIntersectLine(ele: Connectable, line: IVec[]) {
+  if (isTopLevelBlock(ele)) {
+    return linePolygonIntersects(
+      line[0],
+      line[1],
+      Bound.deserialize(ele.xywh).points
+    );
+  } else {
+    return ele.intersectWithLine(line[0], line[1]);
+  }
+}
+
+export function getConnectableNearestAnchor(ele: Connectable, point: IVec) {
+  const anchors = getAnchors(ele);
+  return closestPoint(
+    anchors.map(a => a.point),
+    point
+  );
+}
+
+function closestPoint(points: IVec[], point: IVec) {
+  const rst = points.map(p => ({ p, d: Vec.dist(p, point) }));
+  rst.sort((a, b) => a.d - b.d);
+  return rst[0].p;
+}
+
+function computePoints(
+  startPoint: IVec,
+  endPoint: IVec,
+  nextStartPoint: IVec,
+  lastEndPoint: IVec,
+  startBound: Bound | null,
+  endBound: Bound | null,
+  expandStartBound: Bound | null,
+  expandEndBound: Bound | null
+): [IVec[], IVec, IVec, IVec, IVec] {
+  startPoint = downscalePrecision(startPoint);
+  endPoint = downscalePrecision(endPoint);
+  nextStartPoint = downscalePrecision(nextStartPoint);
+  lastEndPoint = downscalePrecision(lastEndPoint);
+  const result = getConnectablePoints(
+    startPoint,
+    endPoint,
+    nextStartPoint,
+    lastEndPoint,
+    startBound,
+    endBound,
+    expandStartBound,
+    expandEndBound
+  );
+  const points = result.points;
+  nextStartPoint = result.nextStartPoint;
+  lastEndPoint = result.lastEndPoint;
+  const finalPoints = filterConnectablePoints(
+    filterConnectablePoints(points, startBound),
+    endBound
+  );
+  return [finalPoints, startPoint, endPoint, nextStartPoint, lastEndPoint];
+}
+
+function downscalePrecision(point: IVec) {
+  return [
+    Number(point[0].toFixed(2)),
+    Number(point[1].toFixed(2)),
+    point[2] ?? 0,
+  ];
+}
+
+function filterConnectablePoints(points: IVec[], bound: Bound | null) {
+  return points.filter(point => {
+    if (!bound) return true;
+    return !bound.isPointInBound(point);
+  });
+}
+
+function pushWithPriority(points: IVec[], vecs: IVec[], priority = 0) {
+  points.push(...vecs.map(vec => [...vec, priority]));
+}
+
+function pushLineIntersectsToPoints(
+  points: IVec[],
+  aLine: IVec[],
+  bLine: IVec[],
+  priority = 0
+) {
+  const rst = lineIntersects(aLine[0], aLine[1], bLine[0], bLine[1], true);
+  if (rst) {
+    pushWithPriority(points, [rst], priority);
+  }
+}
+
+function pushOuterPoints(
+  points: IVec[],
+  expandStartBound: Bound,
+  expandEndBound: Bound,
+  outerBound: Bound
+) {
+  if (expandStartBound && expandEndBound && outerBound) {
+    pushWithPriority(points, outerBound.getVerticesAndMidpoints());
+    pushWithPriority(points, [outerBound.center], 2);
+    [
+      expandStartBound.upperLine,
+      expandStartBound.horizontalLine,
+      expandStartBound.lowerLine,
+      expandEndBound.upperLine,
+      expandEndBound.horizontalLine,
+      expandEndBound.lowerLine,
+    ].forEach(line => {
+      pushLineIntersectsToPoints(points, line, outerBound.leftLine, 0);
+      pushLineIntersectsToPoints(points, line, outerBound.rightLine, 0);
+    });
+    [
+      expandStartBound.leftLine,
+      expandStartBound.verticalLine,
+      expandStartBound.rightLine,
+      expandEndBound.leftLine,
+      expandEndBound.verticalLine,
+      expandEndBound.rightLine,
+    ].forEach(line => {
+      pushLineIntersectsToPoints(points, line, outerBound.upperLine, 0);
+      pushLineIntersectsToPoints(points, line, outerBound.lowerLine, 0);
+    });
+  }
+}
+
+function pushBoundMidPoint(
+  points: IVec[],
+  bound1: Bound,
+  bound2: Bound,
+  expandBound1: Bound,
+  expandBound2: Bound
+) {
+  if (bound1.maxX < bound2.x) {
+    const midX = (bound1.maxX + bound2.x) / 2;
+    [
+      expandBound1.horizontalLine,
+      expandBound2.horizontalLine,
+      expandBound1.upperLine,
+      expandBound1.lowerLine,
+      expandBound2.upperLine,
+      expandBound2.lowerLine,
+    ].forEach((line, index) => {
+      pushLineIntersectsToPoints(
+        points,
+        line,
+        [
+          [midX, 0],
+          [midX, 1],
+        ],
+        index === 0 || index === 1 ? 6 : 3
+      );
+    });
+  }
+  if (bound1.maxY < bound2.y) {
+    const midY = (bound1.maxY + bound2.y) / 2;
+    [
+      expandBound1.verticalLine,
+      expandBound2.verticalLine,
+      expandBound1.leftLine,
+      expandBound1.rightLine,
+      expandBound2.leftLine,
+      expandBound2.rightLine,
+    ].forEach((line, index) => {
+      pushLineIntersectsToPoints(
+        points,
+        line,
+        [
+          [0, midY],
+          [1, midY],
+        ],
+        index === 0 || index === 1 ? 6 : 3
+      );
+    });
+  }
+}
+
+function pushGapMidPoint(
+  points: IVec[],
+  point: IVec,
+  bound: Bound,
+  bound2: Bound,
+  expandBound: Bound,
+  expandBound2: Bound
+) {
+  /** on top or on bottom */
+  if (
+    almostEqual(point[1], bound.y, 0.02) ||
+    almostEqual(point[1], bound.maxY, 0.02)
+  ) {
+    const rst = [
+      bound.upperLine,
+      bound.lowerLine,
+      bound2.upperLine,
+      bound2.lowerLine,
+    ].map(line => {
+      return lineIntersects(
+        point,
+        [point[0], point[1] + 1],
+        line[0],
+        line[1],
+        true
+      );
+    }) as number[][];
+    rst.sort((a, b) => a[1] - b[1]);
+    const midPoint = Vec.lrp(rst[1], rst[2], 0.5);
+    pushWithPriority(points, [midPoint], 6);
+    [
+      expandBound.leftLine,
+      expandBound.rightLine,
+      expandBound2.leftLine,
+      expandBound2.rightLine,
+    ].forEach(line => {
+      pushLineIntersectsToPoints(
+        points,
+        [midPoint, [midPoint[0] + 1, midPoint[1]]],
+        line,
+        0
+      );
+    });
+  } else {
+    const rst = [
+      bound.leftLine,
+      bound.rightLine,
+      bound2.leftLine,
+      bound2.rightLine,
+    ].map(line => {
+      return lineIntersects(
+        point,
+        [point[0] + 1, point[1]],
+        line[0],
+        line[1],
+        true
+      );
+    }) as number[][];
+    rst.sort((a, b) => a[0] - b[0]);
+    const midPoint = Vec.lrp(rst[1], rst[2], 0.5);
+    pushWithPriority(points, [midPoint], 6);
+    [
+      expandBound.upperLine,
+      expandBound.lowerLine,
+      expandBound2.upperLine,
+      expandBound2.lowerLine,
+    ].forEach(line => {
+      pushLineIntersectsToPoints(
+        points,
+        [midPoint, [midPoint[0], midPoint[1] + 1]],
+        line,
+        0
+      );
+    });
+  }
+}
+
+function removeDulicatePoints(points: IVec[]) {
+  points = points.map(downscalePrecision);
+  points.sort((a, b) => a[0] - b[0]);
+  for (let i = 1; i < points.length - 1; i++) {
+    const cur = points[i];
+    const last = points[i - 1];
+    if (almostEqual(cur[0], last[0], 0.02)) {
+      cur[0] = last[0];
+    }
+  }
+  points.sort((a, b) => a[1] - b[1]);
+  for (let i = 1; i < points.length - 1; i++) {
+    const cur = points[i];
+    const last = points[i - 1];
+    if (almostEqual(cur[1], last[1], 0.02)) {
+      cur[1] = last[1];
+    }
+  }
+  points.sort((a, b) => {
+    if (a[0] < b[0]) return -1;
+    if (a[0] > b[0]) return 1;
+    if (a[1] < b[1]) return -1;
+    if (a[1] > b[1]) return 1;
+    return 0;
+  });
+  for (let i = 1; i < points.length; i++) {
+    const cur = points[i];
+    const last = points[i - 1];
+    if (
+      almostEqual(cur[0], last[0], 0.02) &&
+      almostEqual(cur[1], last[1], 0.02)
+    ) {
+      if (cur[2] <= last[2]) points.splice(i, 1);
+      else points.splice(i - 1, 1);
+      i--;
+      continue;
+    }
+  }
+  return points;
+}
+
+function getConnectablePoints(
+  startPoint: IVec,
+  endPoint: IVec,
+  nextStartPoint: IVec,
+  lastEndPoint: IVec,
+  startBound: Bound | null,
+  endBound: Bound | null,
+  expandStartBound: Bound | null,
+  expandEndBound: Bound | null
+) {
+  const lineBound = Bound.fromPoints([startPoint, endPoint]);
+  const outerBound =
+    expandStartBound &&
+    expandEndBound &&
+    expandStartBound.unite(expandEndBound);
+  let points: IVec[] = [nextStartPoint, lastEndPoint];
+  pushWithPriority(points, lineBound.getVerticesAndMidpoints());
+
+  if (!startBound || !endBound) {
+    pushWithPriority(points, [lineBound.center], 6);
+  }
+  if (outerBound) {
+    pushOuterPoints(points, expandStartBound, expandEndBound, outerBound);
+  }
+
+  if (startBound && endBound && expandStartBound && expandEndBound) {
+    pushGapMidPoint(
+      points,
+      startPoint,
+      startBound,
+      endBound,
+      expandStartBound,
+      expandEndBound
+    );
+    pushGapMidPoint(
+      points,
+      endPoint,
+      endBound,
+      startBound,
+      expandEndBound,
+      expandStartBound
+    );
+    pushBoundMidPoint(
+      points,
+      startBound,
+      endBound,
+      expandStartBound,
+      expandEndBound
+    );
+    pushBoundMidPoint(
+      points,
+      endBound,
+      startBound,
+      expandEndBound,
+      expandStartBound
+    );
+  }
+
+  if (expandStartBound) {
+    pushWithPriority(points, expandStartBound.getVerticesAndMidpoints());
+    pushWithPriority(
+      points,
+      expandStartBound.include(lastEndPoint).getVerticesAndMidpoints()
+    );
+  }
+
+  if (expandEndBound) {
+    pushWithPriority(points, expandEndBound.getVerticesAndMidpoints());
+    pushWithPriority(
+      points,
+      expandEndBound.include(nextStartPoint).getVerticesAndMidpoints()
+    );
+  }
+  points = removeDulicatePoints(points);
+
+  const sorted = points.map(point => point[0] + ',' + point[1]).sort();
+  sorted.forEach((cur, index) => {
+    if (index === 0) return;
+    if (cur === sorted[index - 1]) {
+      throw new Error('duplicate point');
+    }
+  });
+  const startEnds = [nextStartPoint, lastEndPoint].map(point => {
+    return points.find(
+      item =>
+        almostEqual(item[0], point[0], 0.02) &&
+        almostEqual(item[1], point[1], 0.02)
+    );
+  }) as IVec[];
+  assertExists(startEnds[0]);
+  assertExists(startEnds[1]);
+  return { points, nextStartPoint: startEnds[0], lastEndPoint: startEnds[1] };
+}
 
 function getDirectPath(startPoint: IVec, endPoint: IVec) {
   if (
@@ -310,7 +742,7 @@ export class EdgelessConnectorManager {
           position: bound.toRelative(nearestPoint).map(n => clamp(n, 0, 1)),
         };
       }
-
+      if (result) break;
       // if not, check if in original bound
       if (bound.isPointInBound(point)) {
         result = {
